+++ base_layer/key_manager/src/cipher_seed.rs
@@ -20,7 +20,7 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use std::{convert::TryFrom, mem::size_of};
+use std::{convert::TryFrom, mem::size_of, str::FromStr};
 
 use argon2;
 use chacha20::{
@@ -34,12 +34,16 @@ use rand::{rngs::OsRng, RngCore};
 use serde::{Deserialize, Serialize};
 use subtle::ConstantTimeEq;
 use tari_crypto::hash::blake2::Blake256;
+use tari_utilities::{hidden::Hidden, safe_array::SafeArray, SafePassword};
 use zeroize::{Zeroize, Zeroizing};
 
 use crate::{
     error::KeyManagerError,
     mac_domain_hasher,
     mnemonic::{from_bytes, to_bytes, to_bytes_with_language, Mnemonic, MnemonicLanguage},
+    CipherSeedEncryptionKey,
+    CipherSeedMacKey,
+    SeedWords,
     LABEL_ARGON_ENCODING,
     LABEL_CHACHA20_ENCODING,
     LABEL_MAC_GENERATION,
@@ -121,7 +125,7 @@ pub struct CipherSeed {
 }
 
 // This is a separate type to make the linter happy
-type DerivedCipherSeedKeys = Result<(Zeroizing<Vec<u8>>, Zeroizing<Vec<u8>>), KeyManagerError>;
+type DerivedCipherSeedKeys = Result<(CipherSeedEncryptionKey, CipherSeedMacKey), KeyManagerError>;
 
 impl CipherSeed {
     #[cfg(not(target_arch = "wasm32"))]
@@ -165,9 +169,12 @@ impl CipherSeed {
     }
 
     /// Generate an encrypted seed from a passphrase
-    pub fn encipher(&self, passphrase: Option<String>) -> Result<Vec<u8>, KeyManagerError> {
+    pub fn encipher(&self, passphrase: Option<SafePassword>) -> Result<Vec<u8>, KeyManagerError> {
         // Derive encryption and MAC keys from passphrase and main salt
-        let passphrase = Zeroizing::new(passphrase.unwrap_or_else(|| DEFAULT_CIPHER_SEED_PASSPHRASE.to_string()));
+        let passphrase = passphrase.unwrap_or_else(|| {
+            SafePassword::from_str(DEFAULT_CIPHER_SEED_PASSPHRASE)
+                .expect("Failed to parse default cipher seed passphrase to SafePassword")
+        });
         let (encryption_key, mac_key) = Self::derive_keys(&passphrase, self.salt.as_ref())?;
 
         // Generate the MAC
@@ -176,7 +183,7 @@ impl CipherSeed {
             self.entropy.as_ref(),
             CIPHER_SEED_VERSION,
             self.salt.as_ref(),
-            mac_key.as_ref(),
+            &mac_key,
         )?;
 
         // Assemble the secret data to be encrypted: birthday, entropy, MAC
@@ -188,7 +195,7 @@ impl CipherSeed {
         secret_data.extend(&mac);
 
         // Encrypt the secret data
-        Self::apply_stream_cipher(&mut secret_data, encryption_key.as_ref(), self.salt.as_ref())?;
+        Self::apply_stream_cipher(&mut secret_data, &encryption_key, self.salt.as_ref())?;
 
         // Assemble the final seed: version, main salt, secret data, checksum
         let mut encrypted_seed =
@@ -206,7 +213,10 @@ impl CipherSeed {
     }
 
     /// Recover a seed from encrypted data and a passphrase
-    pub fn from_enciphered_bytes(encrypted_seed: &[u8], passphrase: Option<String>) -> Result<Self, KeyManagerError> {
+    pub fn from_enciphered_bytes(
+        encrypted_seed: &[u8],
+        passphrase: Option<SafePassword>,
+    ) -> Result<Self, KeyManagerError> {
         // Check the length: version, birthday, entropy, MAC, salt, checksum
         if encrypted_seed.len() !=
             1 + CIPHER_SEED_BIRTHDAY_BYTES +
@@ -241,7 +251,10 @@ impl CipherSeed {
         }
 
         // Derive encryption and MAC keys from passphrase and main salt
-        let passphrase = Zeroizing::new(passphrase.unwrap_or_else(|| DEFAULT_CIPHER_SEED_PASSPHRASE.to_string()));
+        let passphrase = passphrase.unwrap_or_else(|| {
+            SafePassword::from_str(DEFAULT_CIPHER_SEED_PASSPHRASE)
+                .expect("Failed to parse default cipher seed passphrase to SafePassword")
+        });
         let salt: Box<[u8; CIPHER_SEED_MAIN_SALT_BYTES]> = encrypted_seed
             .split_off(1 + CIPHER_SEED_BIRTHDAY_BYTES + CIPHER_SEED_ENTROPY_BYTES + CIPHER_SEED_MAC_BYTES)
             .into_boxed_slice()
@@ -251,7 +264,7 @@ impl CipherSeed {
 
         // Decrypt the secret data: birthday, entropy, MAC
         let mut secret_data = Zeroizing::new(encrypted_seed.split_off(1));
-        Self::apply_stream_cipher(&mut secret_data, encryption_key.as_ref(), salt.as_ref())?;
+        Self::apply_stream_cipher(&mut secret_data, &encryption_key, salt.as_ref())?;
 
         // Parse secret data
         let mac = secret_data.split_off(CIPHER_SEED_BIRTHDAY_BYTES + CIPHER_SEED_ENTROPY_BYTES);
@@ -266,13 +279,7 @@ impl CipherSeed {
         let birthday = u16::from_le_bytes(birthday_bytes);
 
         // Generate the MAC
-        let expected_mac = Self::generate_mac(
-            &birthday_bytes,
-            entropy.as_ref(),
-            version,
-            salt.as_ref(),
-            mac_key.as_ref(),
-        )?;
+        let expected_mac = Self::generate_mac(&birthday_bytes, entropy.as_ref(), version, salt.as_ref(), &mac_key)?;
 
         // Verify the MAC in constant time to avoid leaking data
         if mac.ct_eq(&expected_mac).unwrap_u8() == 0 {
@@ -288,7 +295,11 @@ impl CipherSeed {
     }
 
     /// Encrypt or decrypt data using ChaCha20
-    fn apply_stream_cipher(data: &mut [u8], encryption_key: &[u8], salt: &[u8]) -> Result<(), KeyManagerError> {
+    fn apply_stream_cipher(
+        data: &mut [u8],
+        encryption_key: &CipherSeedEncryptionKey,
+        salt: &[u8],
+    ) -> Result<(), KeyManagerError> {
         // The ChaCha20 nonce is derived from the main salt
         let encryption_nonce = mac_domain_hasher::<Blake256>(LABEL_CHACHA20_ENCODING)
             .chain(salt)
@@ -296,7 +307,10 @@ impl CipherSeed {
         let encryption_nonce = &encryption_nonce.as_ref()[..size_of::<Nonce>()];
 
         // Encrypt/decrypt the data
-        let mut cipher = ChaCha20::new(Key::from_slice(encryption_key), Nonce::from_slice(encryption_nonce));
+        let mut cipher = ChaCha20::new(
+            Key::from_slice(encryption_key.reveal()),
+            Nonce::from_slice(encryption_nonce),
+        );
         cipher.apply_keystream(data);
 
         Ok(())
@@ -318,7 +332,7 @@ impl CipherSeed {
         entropy: &[u8],
         cipher_seed_version: u8,
         salt: &[u8],
-        mac_key: &[u8],
+        mac_key: &CipherSeedMacKey,
     ) -> Result<Vec<u8>, KeyManagerError> {
         // Check all lengths are valid
         if birthday.len() != CIPHER_SEED_BIRTHDAY_BYTES {
@@ -336,14 +350,14 @@ impl CipherSeed {
             .chain(entropy)
             .chain(&[cipher_seed_version])
             .chain(salt)
-            .chain(mac_key)
+            .chain(mac_key.reveal())
             .finalize()
             .as_ref()[..CIPHER_SEED_MAC_BYTES]
             .to_vec())
     }
 
     /// Use Argon2 to derive encryption and MAC keys from a passphrase and main salt
-    fn derive_keys(passphrase: &str, salt: &[u8]) -> DerivedCipherSeedKeys {
+    fn derive_keys(passphrase: &SafePassword, salt: &[u8]) -> DerivedCipherSeedKeys {
         // The Argon2 salt is derived from the main salt
         let argon2_salt = mac_domain_hasher::<Blake256>(LABEL_ARGON_ENCODING)
             .chain(salt)
@@ -362,15 +376,23 @@ impl CipherSeed {
         .map_err(|_| KeyManagerError::CryptographicError("Problem generating Argon2 parameters".to_string()))?;
 
         // Derive the main key from the password in place
-        let mut main_key = Zeroizing::new([0u8; CIPHER_SEED_ENCRYPTION_KEY_BYTES + CIPHER_SEED_MAC_KEY_BYTES]);
+        let mut main_key = Hidden::hide([0u8; CIPHER_SEED_ENCRYPTION_KEY_BYTES + CIPHER_SEED_MAC_KEY_BYTES]);
         let hasher = argon2::Argon2::new(argon2::Algorithm::Argon2d, argon2::Version::V0x13, params);
         hasher
-            .hash_password_into(passphrase.as_bytes(), argon2_salt, main_key.as_mut())
+            .hash_password_into(passphrase.reveal(), argon2_salt, main_key.reveal_mut())
             .map_err(|_| KeyManagerError::CryptographicError("Problem generating Argon2 password hash".to_string()))?;
 
         // Split off the keys
-        let encryption_key = Zeroizing::new(main_key.as_ref()[..CIPHER_SEED_ENCRYPTION_KEY_BYTES].to_vec());
-        let mac_key = Zeroizing::new(main_key.as_ref()[CIPHER_SEED_ENCRYPTION_KEY_BYTES..].to_vec());
+        let mut encryption_key = CipherSeedEncryptionKey::from(SafeArray::default());
+        encryption_key
+            .reveal_mut()
+            .copy_from_slice(&main_key.reveal()[..CIPHER_SEED_ENCRYPTION_KEY_BYTES]);
+
+        let mut mac_key = CipherSeedMacKey::from(SafeArray::default());
+        mac_key
+            .reveal_mut()
+            .copy_from_slice(&main_key.reveal()[CIPHER_SEED_ENCRYPTION_KEY_BYTES..]);
+
         Ok((encryption_key, mac_key))
     }
 }
@@ -384,34 +406,40 @@ impl Default for CipherSeed {
 impl Mnemonic<CipherSeed> for CipherSeed {
     /// Generates a CipherSeed that represent the provided mnemonic sequence of words, the language of the mnemonic
     /// sequence is autodetected
-    fn from_mnemonic(mnemonic_seq: &[String], passphrase: Option<String>) -> Result<CipherSeed, KeyManagerError> {
+    fn from_mnemonic(
+        mnemonic_seq: &SeedWords,
+        passphrase: Option<SafePassword>,
+    ) -> Result<CipherSeed, KeyManagerError> {
         let bytes = to_bytes(mnemonic_seq)?;
-        CipherSeed::from_enciphered_bytes(&bytes, passphrase)
+        CipherSeed::from_enciphered_bytes(bytes.reveal(), passphrase)
     }
 
     /// Generates a SecretKey that represent the provided mnemonic sequence of words using the specified language
     fn from_mnemonic_with_language(
-        mnemonic_seq: &[String],
+        mnemonic_seq: &SeedWords,
         language: MnemonicLanguage,
-        passphrase: Option<String>,
+        passphrase: Option<SafePassword>,
     ) -> Result<CipherSeed, KeyManagerError> {
         let bytes = to_bytes_with_language(mnemonic_seq, &language)?;
-        CipherSeed::from_enciphered_bytes(&bytes, passphrase)
+        CipherSeed::from_enciphered_bytes(bytes.reveal(), passphrase)
     }
 
     /// Generates a mnemonic sequence of words from the provided secret key
     fn to_mnemonic(
         &self,
         language: MnemonicLanguage,
-        passphrase: Option<String>,
-    ) -> Result<Vec<String>, KeyManagerError> {
+        passphrase: Option<SafePassword>,
+    ) -> Result<SeedWords, KeyManagerError> {
         Ok(from_bytes(&self.encipher(passphrase)?, language)?)
     }
 }
 
 #[cfg(test)]
 mod test {
+    use std::str::FromStr;
+
     use crc32fast::Hasher as CrcHasher;
+    use tari_utilities::{Hidden, SafePassword};
 
     use super::BIRTHDAY_GENESIS_FROM_UNIX_EPOCH;
     use crate::{
@@ -419,27 +447,34 @@ mod test {
         error::KeyManagerError,
         get_birthday_from_unix_epoch_in_seconds,
         mnemonic::{Mnemonic, MnemonicLanguage},
+        SeedWords,
     };
 
     #[test]
     fn test_cipher_seed_generation_and_deciphering() {
         let seed = CipherSeed::new();
 
-        let mut enciphered_seed = seed.encipher(Some("Passphrase".to_string())).unwrap();
+        let mut enciphered_seed = seed
+            .encipher(Some(SafePassword::from_str("Passphrase").unwrap()))
+            .unwrap();
 
         let deciphered_seed =
-            CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("Passphrase".to_string())).unwrap();
+            CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("Passphrase").unwrap()))
+                .unwrap();
 
         assert_eq!(seed, deciphered_seed);
 
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("WrongPassphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(
+            &enciphered_seed,
+            Some(SafePassword::from_str("WrongPassphrase").unwrap()),
+        ) {
             Err(KeyManagerError::DecryptionFailed) => (),
             _ => panic!("Version should not match"),
         }
 
         enciphered_seed[0] = CIPHER_SEED_VERSION + 1; // this is an unsupported version
 
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("Passphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("Passphrase").unwrap())) {
             Err(KeyManagerError::VersionMismatch) => (),
             _ => panic!("Version should not match"),
         }
@@ -449,7 +484,7 @@ mod test {
 
         // flip some bits
         enciphered_seed[1] = !enciphered_seed[1];
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("Passphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("Passphrase").unwrap())) {
             Err(KeyManagerError::CrcError) => (),
             _ => panic!("Crc should not match"),
         }
@@ -475,7 +510,7 @@ mod test {
         enciphered_seed[(n - 4)..].copy_from_slice(&calculated_checksum);
 
         // the MAC decryption should fail in this case
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("passphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("passphrase").unwrap())) {
             Err(KeyManagerError::DecryptionFailed) => (),
             _ => panic!("Decryption should fail"),
         }
@@ -502,7 +537,7 @@ mod test {
         enciphered_seed[(n - 4)..].copy_from_slice(&calculated_checksum);
 
         // the MAC decryption should fail in this case
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("passphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("passphrase").unwrap())) {
             Err(KeyManagerError::DecryptionFailed) => (),
             _ => panic!("Decryption should fail"),
         }
@@ -528,7 +563,7 @@ mod test {
         enciphered_seed[(n - 4)..].copy_from_slice(&calculated_checksum);
 
         // the MAC decryption should fail in this case
-        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some("passphrase".to_string())) {
+        match CipherSeed::from_enciphered_bytes(&enciphered_seed, Some(SafePassword::from_str("passphrase").unwrap())) {
             Err(KeyManagerError::DecryptionFailed) => (),
             _ => panic!("Decryption should fail"),
         }
@@ -559,8 +594,9 @@ mod test {
             "cover", "vote", "federal", "husband", "cave", "alone", "dynamic", "reopen", "visa", "young", "gas",
         ]
         .iter()
-        .map(|x| x.to_string())
-        .collect::<Vec<String>>();
+        .map(|x| Hidden::hide(x.to_string()))
+        .collect::<Vec<Hidden<String>>>();
+        let mnemonic_seq = SeedWords::new(mnemonic_seq);
         // Language not known
         match CipherSeed::from_mnemonic(&mnemonic_seq, None) {
             Ok(_k) => panic!(),
@@ -577,18 +613,24 @@ mod test {
     fn cipher_seed_to_and_from_mnemonic_with_passphrase() {
         let seed = CipherSeed::new();
         let mnemonic_seq = seed
-            .to_mnemonic(MnemonicLanguage::Spanish, Some("Passphrase".to_string()))
+            .to_mnemonic(
+                MnemonicLanguage::Spanish,
+                Some(SafePassword::from_str("Passphrase").unwrap()),
+            )
             .expect("Couldn't convert CipherSeed to Mnemonic");
-        match CipherSeed::from_mnemonic(&mnemonic_seq, Some("Passphrase".to_string())) {
+        match CipherSeed::from_mnemonic(&mnemonic_seq, Some(SafePassword::from_str("Passphrase").unwrap())) {
             Ok(mnemonic_seed) => assert_eq!(seed, mnemonic_seed),
             Err(e) => panic!("Couldn't create CipherSeed from Mnemonic: {}", e),
         }
 
         let mnemonic_seq = seed
-            .to_mnemonic(MnemonicLanguage::Spanish, Some("Passphrase".to_string()))
+            .to_mnemonic(
+                MnemonicLanguage::Spanish,
+                Some(SafePassword::from_str("Passphrase").unwrap()),
+            )
             .expect("Couldn't convert CipherSeed to Mnemonic");
         assert!(
-            CipherSeed::from_mnemonic(&mnemonic_seq, Some("WrongPassphrase".to_string())).is_err(),
+            CipherSeed::from_mnemonic(&mnemonic_seq, Some(SafePassword::from_str("WrongPassphrase").unwrap())).is_err(),
             "Should not be able to derive seed with wrong passphrase"
         );
     }