+++ base_layer/key_manager/src/mnemonic.rs
@@ -24,12 +24,17 @@ use std::{cmp::Ordering, slice::Iter};
 
 use serde::{Deserialize, Serialize};
 use strum_macros::{Display, EnumString};
-use tari_utilities::bit::{bytes_to_bits, checked_bits_to_uint};
+use tari_utilities::{
+    bit::{bytes_to_bits, checked_bits_to_uint},
+    Hidden,
+    SafePassword,
+};
 
 use crate::{
     diacritics::*,
     error::{KeyManagerError, MnemonicError},
     mnemonic_wordlists::*,
+    SeedWords,
 };
 
 /// The Mnemonic system simplifies the encoding and decoding of a secret key into and from a Mnemonic word sequence
@@ -52,7 +57,7 @@ pub enum MnemonicLanguage {
 impl MnemonicLanguage {
     /// Detects the mnemonic language of a specific word by searching all defined mnemonic word lists
     pub fn from(mnemonic_word: &str) -> Result<MnemonicLanguage, MnemonicError> {
-        let words = vec![mnemonic_word.to_string()];
+        let words = SeedWords::new(vec![Hidden::hide(mnemonic_word.to_string())]);
         MnemonicLanguage::detect_language(&words)
     }
 
@@ -84,14 +89,14 @@ impl MnemonicLanguage {
     }
 
     /// Detects the language of a list of words
-    pub fn detect_language(words: &[String]) -> Result<MnemonicLanguage, MnemonicError> {
-        let count = words.iter().len();
+    pub fn detect_language(words: &SeedWords) -> Result<MnemonicLanguage, MnemonicError> {
+        let count = words.len();
         match count.cmp(&1) {
             Ordering::Less => {
                 return Err(MnemonicError::UnknownLanguage);
             },
             Ordering::Equal => {
-                let word = words.get(0).ok_or(MnemonicError::EncodeInvalidLength)?;
+                let word = words.get_word(0)?;
                 for language in MnemonicLanguage::iterator() {
                     if find_mnemonic_index_from_word(word, *language).is_ok() {
                         return Ok(*language);
@@ -100,7 +105,8 @@ impl MnemonicLanguage {
                 return Err(MnemonicError::UnknownLanguage);
             },
             Ordering::Greater => {
-                for word in words {
+                for word_ind in 0..words.len() {
+                    let word = words.get_word(word_ind)?;
                     let mut languages = Vec::with_capacity(MnemonicLanguage::iterator().len());
                     // detect all languages in which a word falls into
                     for language in MnemonicLanguage::iterator() {
@@ -112,7 +118,8 @@ impl MnemonicLanguage {
                     // yielded from the initial word for this iteration
                     for language in languages {
                         let mut consistent = true;
-                        for compare in words {
+                        for compare_ind in 0..words.len() {
+                            let compare = words.get_word(compare_ind)?;
                             if compare != word && find_mnemonic_index_from_word(compare, language).is_err() {
                                 consistent = false;
                             }
@@ -131,19 +138,27 @@ impl MnemonicLanguage {
 
 /// Finds and returns the index of a specific word in a mnemonic word list defined by the specified language
 fn find_mnemonic_index_from_word(word: &str, language: MnemonicLanguage) -> Result<usize, MnemonicError> {
-    let lowercase_word = word.to_lowercase();
+    let lowercase_word = Hidden::hide(word.to_lowercase());
     let search_result = match language {
         // Search through languages are ordered according to the predominance (number of speakers in the world) of that
         // language
         MnemonicLanguage::ChineseSimplified => {
-            MNEMONIC_CHINESE_SIMPLIFIED_WORDS.binary_search(&lowercase_word.as_str())
+            MNEMONIC_CHINESE_SIMPLIFIED_WORDS.binary_search(&lowercase_word.reveal().as_str())
         },
-        MnemonicLanguage::English => MNEMONIC_ENGLISH_WORDS.binary_search(&remove_diacritics(&lowercase_word).as_str()),
-        MnemonicLanguage::French => MNEMONIC_FRENCH_WORDS.binary_search(&remove_diacritics(&lowercase_word).as_str()),
-        MnemonicLanguage::Italian => MNEMONIC_ITALIAN_WORDS.binary_search(&remove_diacritics(&lowercase_word).as_str()),
-        MnemonicLanguage::Japanese => MNEMONIC_JAPANESE_WORDS.binary_search(&lowercase_word.as_str()),
-        MnemonicLanguage::Korean => MNEMONIC_KOREAN_WORDS.binary_search(&lowercase_word.as_str()),
-        MnemonicLanguage::Spanish => MNEMONIC_SPANISH_WORDS.binary_search(&remove_diacritics(&lowercase_word).as_str()),
+        MnemonicLanguage::English => {
+            MNEMONIC_ENGLISH_WORDS.binary_search(&remove_diacritics(lowercase_word.reveal()).reveal().as_str())
+        },
+        MnemonicLanguage::French => {
+            MNEMONIC_FRENCH_WORDS.binary_search(&remove_diacritics(lowercase_word.reveal()).reveal().as_str())
+        },
+        MnemonicLanguage::Italian => {
+            MNEMONIC_ITALIAN_WORDS.binary_search(&remove_diacritics(lowercase_word.reveal()).reveal().as_str())
+        },
+        MnemonicLanguage::Japanese => MNEMONIC_JAPANESE_WORDS.binary_search(&lowercase_word.reveal().as_str()),
+        MnemonicLanguage::Korean => MNEMONIC_KOREAN_WORDS.binary_search(&lowercase_word.reveal().as_str()),
+        MnemonicLanguage::Spanish => {
+            MNEMONIC_SPANISH_WORDS.binary_search(&remove_diacritics(lowercase_word.reveal()).reveal().as_str())
+        },
     };
     match search_result {
         Ok(v) => Ok(v),
@@ -152,54 +167,53 @@ fn find_mnemonic_index_from_word(word: &str, language: MnemonicLanguage) -> Resu
 }
 
 /// Finds and returns the word for a specific index in a mnemonic word list defined by the specified language
-fn find_mnemonic_word_from_index(index: usize, language: MnemonicLanguage) -> Result<String, MnemonicError> {
+fn find_mnemonic_word_from_index(index: usize, language: MnemonicLanguage) -> Result<Hidden<String>, MnemonicError> {
     if index < MNEMONIC_ENGLISH_WORDS.len() {
         Ok(match language {
             // Select word according to specified language
-            MnemonicLanguage::ChineseSimplified => MNEMONIC_CHINESE_SIMPLIFIED_WORDS[index],
-            MnemonicLanguage::English => MNEMONIC_ENGLISH_WORDS[index],
-            MnemonicLanguage::French => MNEMONIC_FRENCH_WORDS[index],
-            MnemonicLanguage::Italian => MNEMONIC_ITALIAN_WORDS[index],
-            MnemonicLanguage::Japanese => MNEMONIC_JAPANESE_WORDS[index],
-            MnemonicLanguage::Korean => MNEMONIC_KOREAN_WORDS[index],
-            MnemonicLanguage::Spanish => MNEMONIC_SPANISH_WORDS[index],
-        }
-        .to_string())
+            MnemonicLanguage::ChineseSimplified => Hidden::hide(MNEMONIC_CHINESE_SIMPLIFIED_WORDS[index].to_string()),
+            MnemonicLanguage::English => Hidden::hide(MNEMONIC_ENGLISH_WORDS[index].to_string()),
+            MnemonicLanguage::French => Hidden::hide(MNEMONIC_FRENCH_WORDS[index].to_string()),
+            MnemonicLanguage::Italian => Hidden::hide(MNEMONIC_ITALIAN_WORDS[index].to_string()),
+            MnemonicLanguage::Japanese => Hidden::hide(MNEMONIC_JAPANESE_WORDS[index].to_string()),
+            MnemonicLanguage::Korean => Hidden::hide(MNEMONIC_KOREAN_WORDS[index].to_string()),
+            MnemonicLanguage::Spanish => Hidden::hide(MNEMONIC_SPANISH_WORDS[index].to_string()),
+        })
     } else {
         Err(MnemonicError::IndexOutOfBounds)
     }
 }
 
 /// Converts a vector of bytes to a sequence of mnemonic words using the specified language
-pub fn from_bytes(bytes: &[u8], language: MnemonicLanguage) -> Result<Vec<String>, MnemonicError> {
-    let mut bits = bytes_to_bits(bytes);
+pub fn from_bytes(bytes: &[u8], language: MnemonicLanguage) -> Result<SeedWords, MnemonicError> {
+    let mut bits = Hidden::hide(bytes_to_bits(bytes));
 
     // Pad with zeros if length not divisible by 11
     let group_bit_count = 11;
-    let mut padded_size = bits.len() / group_bit_count;
-    if bits.len() % group_bit_count > 0 {
+    let mut padded_size = bits.reveal().len() / group_bit_count;
+    if bits.reveal().len() % group_bit_count > 0 {
         padded_size += 1;
     }
     padded_size *= group_bit_count;
-    bits.resize(padded_size, false);
+    bits.reveal_mut().resize(padded_size, false);
 
     // Group each set of 11 bits to form one mnemonic word
-    let mut mnemonic_sequence: Vec<String> = Vec::new();
-    for i in 0..bits.len() / group_bit_count {
+    let mut mnemonic_sequence: Vec<Hidden<String>> = Vec::new();
+    for i in 0..bits.reveal().len() / group_bit_count {
         let start_index = i * group_bit_count;
         let stop_index = start_index + group_bit_count;
-        let sub_v = &bits[start_index..stop_index];
+        let sub_v = &bits.reveal()[start_index..stop_index];
         let word_index = checked_bits_to_uint(sub_v).ok_or(MnemonicError::BitsToIntConversion)?;
         let mnemonic_word = find_mnemonic_word_from_index(word_index, language)?;
         mnemonic_sequence.push(mnemonic_word);
     }
 
-    Ok(mnemonic_sequence)
+    Ok(SeedWords::new(mnemonic_sequence))
 }
 
 /// Generates a vector of bytes that represent the provided mnemonic sequence of words, the language of the mnemonic
 /// sequence is detected
-pub fn to_bytes(mnemonic_seq: &[String]) -> Result<Vec<u8>, MnemonicError> {
+pub fn to_bytes(mnemonic_seq: &SeedWords) -> Result<Hidden<Vec<u8>>, MnemonicError> {
     let language = MnemonicLanguage::detect_language(mnemonic_seq)?;
     to_bytes_with_language(mnemonic_seq, &language)
 }
@@ -219,43 +233,51 @@ pub fn to_bytes(mnemonic_seq: &[String]) -> Result<Vec<u8>, MnemonicError> {
 /// 1) the first output 'a' is last 8 bits from input 'A', we have leftover 3 bits from 'A'
 /// 2) We add 5 bits from 'B' to generate 'b', the leftover is 6 bits from 'B'
 /// 3) We add 2 bits from 'C to generate 'c', now we have 8 bits needed to generate 'd' and we have 1 bit leftover.
-pub fn to_bytes_with_language(mnemonic_seq: &[String], language: &MnemonicLanguage) -> Result<Vec<u8>, MnemonicError> {
+pub fn to_bytes_with_language(
+    mnemonic_seq: &SeedWords,
+    language: &MnemonicLanguage,
+) -> Result<Hidden<Vec<u8>>, MnemonicError> {
     const MASK: u64 = (1u64 << 8) - 1;
-    let mut bytes = Vec::new();
+    let mut bytes = Hidden::hide(Vec::new());
     let mut rest = 0u64;
     let mut rest_bits: u8 = 0;
 
-    for curr_word in mnemonic_seq {
-        let index = find_mnemonic_index_from_word(curr_word, *language)? as u64;
+    for curr_ind in 0..mnemonic_seq.len() {
+        let index = find_mnemonic_index_from_word(
+            mnemonic_seq
+                .get_word(curr_ind)
+                .map_err(|_| MnemonicError::IndexOutOfBounds)?,
+            *language,
+        )? as u64;
         // Add 11 bits to the front
         rest += index << rest_bits;
         rest_bits += 11;
         while rest_bits >= 8 {
             // Get last 8 bits and shift it
-            bytes.push((rest & MASK) as u8);
+            bytes.reveal_mut().push((rest & MASK) as u8);
             rest >>= 8;
             rest_bits -= 8;
         }
     }
     // If we have any leftover, we write it.
     if rest > 0 {
-        bytes.push((rest & MASK) as u8);
+        bytes.reveal_mut().push((rest & MASK) as u8);
     }
     Ok(bytes)
 }
 
 pub trait Mnemonic<T> {
-    fn from_mnemonic(mnemonic_seq: &[String], passphrase: Option<String>) -> Result<T, KeyManagerError>;
+    fn from_mnemonic(mnemonic_seq: &SeedWords, passphrase: Option<SafePassword>) -> Result<T, KeyManagerError>;
     fn from_mnemonic_with_language(
-        mnemonic_seq: &[String],
+        mnemonic_seq: &SeedWords,
         language: MnemonicLanguage,
-        passphrase: Option<String>,
+        passphrase: Option<SafePassword>,
     ) -> Result<T, KeyManagerError>;
     fn to_mnemonic(
         &self,
         language: MnemonicLanguage,
-        passphrase: Option<String>,
-    ) -> Result<Vec<String>, KeyManagerError>;
+        passphrase: Option<SafePassword>,
+    ) -> Result<SeedWords, KeyManagerError>;
 }
 
 #[cfg(test)]
@@ -335,45 +357,54 @@ mod test {
         assert!(MnemonicLanguage::from("desvelado").is_err()); // Invalid Mnemonic Spanish word
 
         // English/Spanish + English/French -> English
-        let words1 = vec![
-            "album".to_string(),
-            "area".to_string(),
-            "opera".to_string(),
-            "abandon".to_string(),
-        ];
+        let words1 = SeedWords::new(vec![
+            Hidden::hide("album".to_string()),
+            Hidden::hide("area".to_string()),
+            Hidden::hide("opera".to_string()),
+            Hidden::hide("abandon".to_string()),
+        ]);
         assert_eq!(
             MnemonicLanguage::detect_language(&words1),
             Ok(MnemonicLanguage::English)
         );
 
         // English/Spanish + English/French + Italian/Spanish
-        let words2 = vec![
-            "album".to_string(),
-            "area".to_string(),
-            "opera".to_string(),
-            "abandon".to_string(),
-            "tipico".to_string(),
-        ];
+        let words2 = SeedWords::new(vec![
+            Hidden::hide("album".to_string()),
+            Hidden::hide("area".to_string()),
+            Hidden::hide("opera".to_string()),
+            Hidden::hide("abandon".to_string()),
+            Hidden::hide("tipico".to_string()),
+        ]);
         assert!(MnemonicLanguage::detect_language(&words2).is_err());
 
         // bounds check (last word is invalid)
-        let words3 = vec![
-            "album".to_string(),
-            "area".to_string(),
-            "opera".to_string(),
-            "abandon".to_string(),
-            "topazio".to_string(),
-        ];
+        let words3 = SeedWords::new(vec![
+            Hidden::hide("album".to_string()),
+            Hidden::hide("area".to_string()),
+            Hidden::hide("opera".to_string()),
+            Hidden::hide("abandon".to_string()),
+            Hidden::hide("topazio".to_string()),
+        ]);
         assert!(MnemonicLanguage::detect_language(&words3).is_err());
 
         // building up a word list: English/French + French -> French
         let mut words = Vec::with_capacity(3);
-        words.push("concert".to_string());
-        assert_eq!(MnemonicLanguage::detect_language(&words), Ok(MnemonicLanguage::English));
-        words.push("abandon".to_string());
-        assert_eq!(MnemonicLanguage::detect_language(&words), Ok(MnemonicLanguage::English));
-        words.push("barbier".to_string());
-        assert_eq!(MnemonicLanguage::detect_language(&words), Ok(MnemonicLanguage::French));
+        words.push(Hidden::hide("concert".to_string()));
+        assert_eq!(
+            MnemonicLanguage::detect_language(&SeedWords::new(words.clone())),
+            Ok(MnemonicLanguage::English)
+        );
+        words.push(Hidden::hide("abandon".to_string()));
+        assert_eq!(
+            MnemonicLanguage::detect_language(&SeedWords::new(words.clone())),
+            Ok(MnemonicLanguage::English)
+        );
+        words.push(Hidden::hide("barbier".to_string()));
+        assert_eq!(
+            MnemonicLanguage::detect_language(&SeedWords::new(words)),
+            Ok(MnemonicLanguage::French)
+        );
     }
 
     #[test]
@@ -384,7 +415,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::ChineseSimplified).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::ChineseSimplified).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using English Simplified
         let desired_index = 1717;
@@ -392,7 +423,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::English).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::English).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using French Simplified
         let desired_index = 824;
@@ -400,7 +431,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::French).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::French).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using Italian Simplified
         let desired_index = 1123;
@@ -408,7 +439,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::Italian).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::Italian).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using Japanese Simplified
         let desired_index = 1856;
@@ -416,7 +447,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::Japanese).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::Japanese).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using Korean Simplified
         let desired_index = 345;
@@ -424,7 +455,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::Korean).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::Korean).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
 
         // Encoding and Decoding using Spanish Simplified
         let desired_index = 345;
@@ -432,7 +463,7 @@ mod test {
         let index = find_mnemonic_index_from_word(&desired_word, MnemonicLanguage::Spanish).expect("");
         assert_eq!(desired_index, index);
         let word = find_mnemonic_word_from_index(desired_index, MnemonicLanguage::Spanish).expect("");
-        assert_eq!(desired_word, word);
+        assert_eq!(&desired_word, word.reveal());
     }
 
     #[test]
@@ -442,7 +473,7 @@ mod test {
         let mnemonic_bytes = mnemonic::to_bytes(&mnemonic_seq).expect("");
         let mismatched_bytes = secretkey_bytes
             .iter()
-            .zip(mnemonic_bytes.iter())
+            .zip(mnemonic_bytes.reveal().iter())
             .filter(|&(a, b)| a != b)
             .count();
         assert_eq!(mismatched_bytes, 0);
@@ -458,7 +489,7 @@ mod test {
             OsRng.fill_bytes(&mut secretkey_bytes);
             let mnemonic_seq = mnemonic::from_bytes(&secretkey_bytes, MnemonicLanguage::English).unwrap();
             let mnemonic_bytes = mnemonic::to_bytes(&mnemonic_seq).unwrap();
-            assert_eq!(secretkey_bytes, mnemonic_bytes, "failed len = {}", len);
+            assert_eq!(&secretkey_bytes, mnemonic_bytes.reveal(), "failed len = {}", len);
         }
     }
 }